todo_list/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0